chore: move protocol package into P2P module
(cherry picked from commit 655e61a9595ee20a2edb35520b05f8ee62221b39)
